Refactoring entities reattempt
